Authorize and fix total balance operation

+++ AlkemyMiniChallenge/Controllers/CategoriesController.cs
@@ -8,27 +8,34 @@ using Microsoft.EntityFrameworkCore;
 using AlkemyMiniChallenge.Data;
 using AlkemyMiniChallenge.Models;
 using AlkemyMiniChallenge.Services;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Authorization;
 
 namespace AlkemyMiniChallenge.Controllers
 {
     public class CategoriesController : Controller
     {
         private readonly ApplicationDbContext _context;
+        private readonly UserManager<IdentityUser> _userManager;
 
-        public CategoriesController(ApplicationDbContext context)
+        public CategoriesController(ApplicationDbContext context, UserManager<IdentityUser> userManager)
         {
+            _userManager = userManager;
             _context = context;
         }
 
         // GET: Categories
+        [Authorize]
         public async Task<IActionResult> Index(
         string sortOrder,
         string currentFilter,
         int? pageNumber)
         {
             ViewData["NameSortParm"] = String.IsNullOrEmpty(sortOrder) ? "name_desc" : "";
-            var category = from s in _context.Category
-                           select s;
+            
+            var userId = _userManager.GetUserId(User);
+            var category = from s in _context.Category.Where(o => o.UserId == userId)
+                            select s;
 
             category = sortOrder switch
             {
@@ -40,6 +47,7 @@ namespace AlkemyMiniChallenge.Controllers
         }
 
         // GET: Categories/Details/5
+        [Authorize]
         public async Task<IActionResult> Details(int? id)
         {
             if (id == null)
@@ -47,17 +55,24 @@ namespace AlkemyMiniChallenge.Controllers
                 return NotFound();
             }
 
-            var category = await _context.Category
-                .FirstOrDefaultAsync(m => m.Id == id);
+            var currentUser = await _userManager.GetUserAsync(HttpContext.User);
+            var category = await _context.Category.FirstOrDefaultAsync(m => m.Id == id);
+
             if (category == null)
             {
                 return NotFound();
+            }
+            
+            if (category.UserId != currentUser.Id)
+            {
+                return RedirectToAction(nameof(Index));
             }
 
             return View(category);
         }
 
         // GET: Categories/Create
+        [Authorize]
         public IActionResult Create()
         {
             return View();
@@ -70,6 +85,9 @@ namespace AlkemyMiniChallenge.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,Name")] Category category)
         {
+            var userId = _userManager.GetUserId(User);
+            category.UserId = userId;
+
             if (ModelState.IsValid && CategoryNameExists(category.Name)!=true)
             {
                 _context.Add(category);
@@ -80,6 +98,7 @@ namespace AlkemyMiniChallenge.Controllers
         }
 
         // GET: Categories/Edit/5
+        [Authorize]
         public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
@@ -87,11 +106,20 @@ namespace AlkemyMiniChallenge.Controllers
                 return NotFound();
             }
 
+            
+            var currentUser = await _userManager.GetUserAsync(HttpContext.User);
             var category = await _context.Category.FindAsync(id);
+
             if (category == null)
             {
                 return NotFound();
             }
+
+            if (category.UserId != currentUser.Id)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+
             return View(category);
         }
 
@@ -100,11 +128,20 @@ namespace AlkemyMiniChallenge.Controllers
         // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Authorize]
         public async Task<IActionResult> Edit(int id, [Bind("Id,Name")] Category category)
         {
             if (id != category.Id)
             {
                 return NotFound();
+            }
+
+            var currentUser = await _userManager.GetUserAsync(HttpContext.User);            
+            var existingCategory = await _context.Category.FirstOrDefaultAsync(m => m.Id == id);
+
+            if (existingCategory.UserId != currentUser.Id)
+            {
+                return RedirectToAction(nameof(Index));
             }
 
             if (ModelState.IsValid && CategoryNameExists(category.Name) != true)
@@ -131,18 +168,25 @@ namespace AlkemyMiniChallenge.Controllers
         }
 
         // GET: Categories/Delete/5
+        [Authorize]
         public async Task<IActionResult> Delete(int? id)
         {
             if (id == null)
             {
                 return NotFound();
             }
-
-            var category = await _context.Category
-                .FirstOrDefaultAsync(m => m.Id == id);
+            
+            var currentUser = await _userManager.GetUserAsync(HttpContext.User);
+            var category = await _context.Category.FirstOrDefaultAsync(m => m.Id == id);
+            
             if (category == null)
             {
                 return NotFound();
+            }
+
+            if (category.UserId != currentUser.Id)
+            {
+                return RedirectToAction(nameof(Index));
             }
 
             return View(category);
@@ -151,10 +195,19 @@ namespace AlkemyMiniChallenge.Controllers
         // POST: Categories/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
+        [Authorize]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            var currentUser = await _userManager.GetUserAsync(HttpContext.User);
             var category = await _context.Category.FindAsync(id);
+
+            if (category.UserId != currentUser.Id)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+           
             _context.Category.Remove(category);
+
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
